Fixed C++ project issue when not placed in Marketplace folder. Fixed 5.5 deprecated warnings.
In a C++ Unreal Engine project, when this plugin is installed to the Engine and NOT in the plugin's Marketplace folder you are met with this error "Expecting to find a type to be declared in a module rules named ‘GaeaUnrealTools’ in UE5Rules". If you put this plugin in the Marketplace folder then the Icon is missing from the Editor button. This has been resolved.
Updated the editor code to the new API to resolve deprecation warnings.